requirements.txt
keyboard==0.13.5
pyautogui==0.9.53
Pillow~=9.1.0